refactor(precompile): create crypto submodule for secp256k1
Move `ecrecover_bytes` and the backend selection (`cfg_if!` for
k256 vs bitcoin_secp256k1) into a dedicated `crypto.rs` submodule,
separating pure cryptographic logic from the precompile runner.

Closes part of #2782
